fix: add missing assertion method
