+++ ArkEcosystem.Client.Tests/Fixtures/wallets-dummy-locks.json
@@ -5,9 +5,9 @@
     "totalCount": 1,
     "next": null,
     "previous": null,
-    "self": "/api/wallets/DAWNRvnMNPWodLjQwhphJwUpTW2KJXBzVR/locks?page=1&limit=100",
-    "first": "/api/wallets/DAWNRvnMNPWodLjQwhphJwUpTW2KJXBzVR/locks?page=1&limit=100",
-    "last": "/api/wallets/DAWNRvnMNPWodLjQwhphJwUpTW2KJXBzVR/locks?page=1&limit=100"
+    "self": "/api/wallets/dummy/locks?page=1&limit=100",
+    "first": "/api/wallets/dummy/locks?page=1&limit=100",
+    "last": "/api/wallets/dummy/locks?page=1&limit=100"
   },
   "data": [
     {
